AI/requirements.txt
fastapi==0.94.1
uvicorn==0.21.0
pytz==2021.3
boto3==1.26.90
python-dotenv==1.0.0
pillow==9.4.0